backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Stack
torch==2.9.0
torchvision==0.19.0
open-clip-torch==2.23.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
boto3==1.34.0

# Performance & Caching
redis==4.5.4
aioredis==2.0.1

# Security & Validation
python-magic==0.4.27
slowapi==0.1.9

# Monitoring & Logging
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0